refactor: fold coverage-area helper into _summary, merge CustomStatistic into AggOperations
- drop the unused calculate_patrol_coverage task; _coverage_area_km2 now
  lives beside its only caller in _summary.py (area_crs defaults to EPSG:6933)
- trim AggOperations to plain pandas aggregations; night_day_ratio and
  coverage_area belong to their SummaryParam variants, and aggregate_over_rows
  / ecoplot agg fields no longer offer options that would crash df.agg
- CustomMetric's Statistic dropdown now reuses AggOperations
- tighten CoverageSummaryParam.swath_width_meters to plain float, removing
  the dead 'or 500.0' fallback
