+++ ecoscope/platform/tasks/analysis/__init__.py
@@ -18,7 +18,6 @@ from ._coverage_weighting import (
     set_coverage_weighting,
 )
 from ._create_meshgrid import create_meshgrid
-from ._patrol_coverage import calculate_patrol_coverage
 from ._patrol_summary import set_patrol_summary_metrics
 from ._summary import aggregate_over_rows, summarize_df
 from ._time_density import (
@@ -45,7 +44,6 @@ __all__ = [
     "normalize_coverage_density",
     "set_coverage_weighting",
     "create_meshgrid",
-    "calculate_patrol_coverage",
     "aggregate_over_rows",
     "set_patrol_summary_metrics",
     "summarize_df",
