+++ tests/platform/tasks/test_summary.py
@@ -11,6 +11,7 @@ from ecoscope.platform.tasks.analysis._summary import (
     NightDayRatioSummaryParam,
     NumericSummaryParam,
     TallySummaryParam,
+    _coverage_area_km2,
     summarize_df,
 )
 
@@ -145,6 +146,11 @@ def test_summarize_df_coverage_scales_with_swath(coverage_trajectories):
     # Buffered-line area is dominated by length * width, so doubling the swath
     # roughly doubles the covered area.
     assert unmerged(1000.0) == pytest.approx(2 * unmerged(500.0), rel=0.05)
+
+
+def test_coverage_area_km2_empty():
+    empty = gpd.GeoDataFrame({"geometry": []}, crs="EPSG:4326")
+    assert _coverage_area_km2(empty, 500.0, merged=True) == 0.0
 
 
 def test_summarize_df_coverage_groupby(coverage_trajectories):
